Loose Trigger includes all peds now
